refactor(api): remove `api.py`; add api folder and modularize api scripts
