+++ montage/api/__init__.py
@@ -1,0 +1,8 @@
+from .document import DocumentAPI
+from .file import FileAPI
+from .policy import PolicyAPI
+from .role import RoleAPI
+from .schema import SchemaAPI
+from .user import UserAPI
+
+__all__ = ('DocumentAPI', 'FileAPI', 'PolicyAPI', 'RoleAPI', 'SchemaAPI', 'UserAPI')
